+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Videos with changed fps from fps_changer will be stored in <TMP_DIR>/fps_changer.
 - Selector now case-insensitive, so it selects mp4 and MP4 etc.
 
 ## 🎉 [0.3.2] - 2024-01-14